Replace SubDAGs with TaskGroups.
