+++ dags/mozaggregator_mobile.py
@@ -12,7 +12,6 @@ import os
 from datetime import datetime, timedelta
 
 from airflow import DAG
-from airflow.operators.subdag import SubDagOperator
 from airflow.providers.google.cloud.operators.gcs import GCSDeleteObjectsOperator
 from utils.dataproc import copy_artifacts_dev, moz_dataproc_pyspark_runner
 from utils.gcp import gke_command
@@ -45,8 +44,8 @@ dag = DAG(
     tags=tags,
 )
 
-subdag_args = default_args.copy()
-subdag_args["retries"] = 0
+task_group_args = default_args.copy()
+task_group_args["retries"] = 0
 
 task_id = "mobile_aggregate_view_dataproc"
 gcp_conn_id = "google_cloud_airflow_dataproc"
@@ -70,55 +69,51 @@ storage_bucket = (
 output_bucket = artifact_bucket if is_dev else "airflow-dataproc-bq-parquet-exports"
 
 
-mobile_aggregate_view_dataproc = SubDagOperator(
-    task_id=task_id,
+mobile_aggregate_view_dataproc = moz_dataproc_pyspark_runner(
+    task_group_name=task_id,
     dag=dag,
-    subdag=moz_dataproc_pyspark_runner(
-        parent_dag_name=dag.dag_id,
-        dag_name=task_id,
-        job_name="mobile_aggregates",
-        cluster_name="mobile-metrics-aggregates-{{ ds_nodash }}",
-        idle_delete_ttl=600,
-        num_workers=5,
-        worker_machine_type="n1-standard-8",
-        init_actions_uris=[
-            "gs://dataproc-initialization-actions/python/pip-install.sh"
-        ],
-        additional_properties={
-            "spark:spark.jars": "gs://spark-lib/bigquery/spark-bigquery-latest.jar",
-            "spark:spark.jars.packages": "org.apache.spark:spark-avro_2.11:2.4.4",
-        },
-        additional_metadata={
-            "PIP_PACKAGES": "git+https://github.com/mozilla/python_mozaggregator.git@pbd_fix_2"
-        },
-        python_driver_code="gs://{}/jobs/mozaggregator_runner.py".format(
-            artifact_bucket
-        ),
-        py_args=[
-            "mobile",
-            "--date",
-            "{{ ds_nodash }}",
-            "--output",
-            f"gs://{output_bucket}/mobile_metrics_aggregates/v3",
-            "--num-partitions",
-            str(5 * 32),
-        ]
-        + (
-            ["--source", "bigquery", "--project-id", "moz-fx-data-shared-prod"]
-            if not EXPORT_TO_AVRO
-            else [
-                "--source",
-                "avro",
-                "--avro-prefix",
-                "gs://airflow-dataproc-bq-parquet-exports-tmp/avro/mozaggregator/mobile/moz-fx-data-shared-prod",
-            ]
-        ),
-        gcp_conn_id=gcp_conn_id,
-        service_account=client_email,
-        artifact_bucket=artifact_bucket,
-        storage_bucket=storage_bucket,
-        default_args=subdag_args,
+    job_name="mobile_aggregates",
+    cluster_name="mobile-metrics-aggregates-{{ ds_nodash }}",
+    idle_delete_ttl=600,
+    num_workers=5,
+    worker_machine_type="n1-standard-8",
+    init_actions_uris=[
+        "gs://dataproc-initialization-actions/python/pip-install.sh"
+    ],
+    additional_properties={
+        "spark:spark.jars": "gs://spark-lib/bigquery/spark-bigquery-latest.jar",
+        "spark:spark.jars.packages": "org.apache.spark:spark-avro_2.11:2.4.4",
+    },
+    additional_metadata={
+        "PIP_PACKAGES": "git+https://github.com/mozilla/python_mozaggregator.git@pbd_fix_2"
+    },
+    python_driver_code="gs://{}/jobs/mozaggregator_runner.py".format(
+        artifact_bucket
     ),
+    py_args=[
+        "mobile",
+        "--date",
+        "{{ ds_nodash }}",
+        "--output",
+        f"gs://{output_bucket}/mobile_metrics_aggregates/v3",
+        "--num-partitions",
+        str(5 * 32),
+    ]
+    + (
+        ["--source", "bigquery", "--project-id", "moz-fx-data-shared-prod"]
+        if not EXPORT_TO_AVRO
+        else [
+            "--source",
+            "avro",
+            "--avro-prefix",
+            "gs://airflow-dataproc-bq-parquet-exports-tmp/avro/mozaggregator/mobile/moz-fx-data-shared-prod",
+        ]
+    ),
+    gcp_conn_id=gcp_conn_id,
+    service_account=client_email,
+    artifact_bucket=artifact_bucket,
+    storage_bucket=storage_bucket,
+    default_args=task_group_args,
 )
 
 # export to avro, if necessary
