+++ utils/glam/generate_query.py
